+++ packages/finschia/src/modules/token/queries.spec.ts
@@ -2,7 +2,7 @@ import { coins, DirectSecp256k1HdWallet } from "@cosmjs/proto-signing";
 import { assertIsDeliverTxSuccess, QueryClient } from "@cosmjs/stargate";
 import { Tendermint34Client } from "@cosmjs/tendermint-rpc";
 import { assert, sleep } from "@cosmjs/utils";
-import { Permission } from "lbmjs-types/lbm/token/v1/token";
+import { Permission } from "@finschia/finschia-proto/lbm/token/v1/token";
 
 import { makeLinkPath } from "../../paths";
 import { SigningFinschiaClient } from "../../signingfinschiaclient";